Users/Matia/Downloads/DjangoReactJWT-main/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
PyJWT==2.9.0
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2
